+++ src/template_augmentation/console_fetcher.rs
@@ -1,4 +1,4 @@
-use super::{KeyExtact, OptAugmentationResult};
+use super::{template_extractation::ExtractForConsole, OptAugmentationResult};
 
 mod test_console_fetcher;
 pub use test_console_fetcher::TestConsoleFetcher;
@@ -7,5 +7,5 @@ mod io_console_fetcher;
 pub use io_console_fetcher::IoConsoleFetcher;
 
 pub trait ConsoleFetcher {
-    fn fetch_from(&self, key: KeyExtact<'_>) -> OptAugmentationResult;
+    fn fetch_from(&self, extract: &ExtractForConsole) -> OptAugmentationResult;
 }