+++ src/template_augmentation/console_fetcher/test_console_fetcher.rs
@@ -1,6 +1,8 @@
 use std::{cell::RefCell, collections::HashSet};
 
-use crate::template_augmentation::{AugementKey, FakeAugmentStore, KeyExtact, TestAugmentStore};
+use crate::template_augmentation::{
+    template_extractation::ExtractForConsole, AugementKey, FakeAugmentStore, TestAugmentStore,
+};
 
 use super::ConsoleFetcher;
 
@@ -10,14 +12,18 @@ pub struct TestConsoleFetcher {
 }
 
 impl ConsoleFetcher for TestConsoleFetcher {
-    fn fetch_from(&self, key: KeyExtact) -> crate::template_augmentation::OptAugmentationResult {
+    fn fetch_from(
+        &self,
+        extract: &ExtractForConsole<'_>,
+    ) -> crate::template_augmentation::OptAugmentationResult {
+        let ExtractForConsole { key, .. } = extract;
         let mut mut_store = self.store.borrow_mut();
         let mut mut_founds = self.accessed_keys.borrow_mut();
         if !mut_founds.insert(key.to_string()) {
             panic!("Key: {} already accessed", key);
         }
 
-        match mut_store.get_mut(key) {
+        match mut_store.get_mut(*key) {
             None => Ok(None),
             Some(value) => Ok(Some(value.clone())),
         }