+++ src/template_augmentation/template_extractation.rs
@@ -2,8 +2,17 @@ use super::{DefaultExtact, KeyExtact};
 
 #[derive(Debug)]
 pub enum TemplateExtractation<'t> {
-    FromConsole {
-        key: KeyExtact<'t>,
-        default_value: DefaultExtact<'t>,
-    },
+    FromConsole(ExtractForConsole<'t>),
+}
+
+#[derive(Debug)]
+pub struct ExtractForConsole<'a> {
+    pub key: KeyExtact<'a>,
+    pub default_value: DefaultExtact<'a>,
+}
+
+impl<'a> ExtractForConsole<'a> {
+    pub fn new(key: KeyExtact<'a>, default_value: DefaultExtact<'a>) -> Self {
+        Self { key, default_value }
+    }
 }